changed error codes in deleteProduct

+++ backend/controllers/product.controller.js
@@ -31,12 +31,17 @@ export const createProduct = async (req,res) => {
 
 export const deleteProduct = async (req, res) => { // :id is placeholder for id
     const { id } = req.params; // id is from :id
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message:"Product deleted" });
     } catch (error) {
         console.log("error in fetching producs:", error.message);
-        res.status(404).json({ success:false, message:"Product not found"});
+        res.status(500).json({ success:false, message:"Server error"});
     }
 };
